Enable static, bundled libcurl by default
Sometimes when using the platform provided libcurl there can be subtle differences between systems that can be hard to track down. Unless someone needs to use the platform libcurl specifically, we will now use the bundled libcurl by default. This will result in more consistent behavior between platforms that is easier to maintain.

The static libcurl is now behind the `static-curl` feature (enabled by default) so it can be disabled if needed.