fix: Ensure important code runs in release mode
Previously, certain critical lines of code were accidentally wrapped in `debug_assert!` when `assert!` would have been appropriate. It seems that I misunderstood the implications of `debug_assert!`; the end result was that a critical piece of code, the code that resolves the request future, was put inside one such statement, resulting in it never being run in release mode.

Also, add some code to ensure that dropping a `Client` is managed correctly.