fix: Ensure important code runs in release mode (#32)
Previously, certain critical lines of code were accidentally wrapped in `debug_assert!` when `assert!` would have been appropriate. It seems that I misunderstood the implications of `debug_assert!`; the end result was that a critical piece of code, the code that resolves the request future, was put inside one such statement, resulting in it never being run in release mode.

This issue manifested itself in the internal future for a request never getting resolved before being closed, resulting in the caller getting a `Canceled` error.

- Add some code to ensure that dropping a `Client` is managed correctly.
- Run an example compiled in release mode during builds to ensure problems like these are caught before being merged.

Fixes #30.