fix: Refresh PSL during run time
The existing means that we use the Public Suffix List has a number of problems:

- Requires connecting to the Internet during builds.
- Slow compilation times (sometimes because of the above).
- The list is fixed at compile time, and cannot be updated without recompiling software. This is bad news for long-running applications that stay in production for a long time.

Switch to the [publicsuffix](https://docs.rs/publicsuffix) crate (by the same author as the psl crate) which does things at runtime instead of compile time. We include a default list at compile time still, but a pinned offline copy instead of downloaded. Also refresh the list from the Internet periodically in the background to avoid potential privacy problems.

- [ ] TODO: Don't spawn additional refresh threads if one is already started.
- [ ] TODO: If we can't refresh the list, try again after a delay to prevent spamming with attempts.