fix: use CURLOPT_POSTFIELDSIZE for POST body
When sending a POST request with a body, the body length must be specified using CURLOPT_POSTFIELDSIZE, not CURLOPT_INFILESIZE like for other request methods.

If the user provides a `Content-Length` header, use that as the request body size, even if the body itself has an unknown size.

In addition, for bodies with an unknown size, set `Transfer-Encoding: chunked` explicitly to ensure that curl knows to use chunked encoding.

Add more comprehensive tests for sending a request body to ensure that such bugs are caught in the future.

Fixes #51.