Add ResponseExt::effective_uri
Implement a new response extension method, `effective_uri`, which returns the last URI that was requested on behalf of a single transfer. This is the same as the URI provided if redirect following is disabled, otherwise the last URI followed.

Implemented using the raw curl API since this is not currently possible in the safe wrapper API.

Fix the cookie jar middleware to use the new `effective_uri` method for computing cookie allowed domains. Previously this panicked, as it was attempting to access a response extension that was never implemented.

Fixes #70.