Add options for configuring the connection cache
There are good use cases why you might want to limit the size of the connection cache, or disable it altogether.

Add `connection_cache_size` to `HttpClientBuilder`, which allows you to override the defaults and set a maximum number of persistent connections to keep in the cache, or to disable connection caching altogether.

In addition, add `disable_connection_cache` to `RequestBuilderExt`, which allows you to opt-out of connection caching on a per-request basis.

Fixes part of #79.