Add new VersionNegotiation API (#129)
Replace `preferred_http_version()` with a new `VersionNegotiation` API that provides more explicit and clear settings for how HTTP versions will be negotiated. This also provides support for "HTTP/2 Prior Knowledge", and will allow us to support HTTP/3 when it comes in the future in a backwards-compatible way.

This is implemented almost exactly as proposed in #108.

Fixes #108.