Make default VersionNegotiation more conservative
The `Upgrade` header that can be used for insecure upgrade from HTTP/1.x to HTTP/2 is not handled properly by many in-production HTTP servers and can result in an error when the request would otherwise succeed without the `Upgrade` header.

Since there's not really any way for us to know if an error response was caused by our `Upgrade` header or not (and what the server state is for non-idempotent methods), `VersionNegotiation::latest_compatible` will no longer send this header, and instead rely exclusively on the server to explicitly _declare_ higher HTTP versions that it supports, either via ALPN or something else.

Fixes #159.