Replace default_headers_mut with default_headers (#186)
Thinking about future compatibility, we might want to leverage interceptors someday as an implementation detail for how default headers are set. To avoid locking us in the current implementation, do not expose the default `HeaderMap` as a direct reference. Instead, for users who want to set headers in bulk, provide `default_headers` which takes an iterator of key-value pairs and replaces all existing headers. That allows users to do the same kind of things without relying on a mutable `HeaderMap` being exposed.

Also tweak some docs.