Fix HEAD requests with a body on libcurl 7.70+ (#213)
libcurl 7.71.0 introduced a breaking change that caused the order in which `CURLOPT_NOBODY` and `CURLOPT_UPLOAD` are set to matter. Previously we were setting `CURLOPT_UPLOAD` first and then `CURLOPT_NOBODY` when the user wishes to make a HEAD request with a request body.

Change the order of operations such that the response body is still ignored, but the internal request type is still PUT.

Fixes #214.