Add socket dial configuration (#207)
This adds a new type of configuration, "dial", which allows you to control and override how Isahc connects to a host. Initially, this will allow you to specify a UNIX socket file to connect to, but this will also be expanded to allow you to specify custom IP socket addresses as well.

This interface also leaves room to add the ability for users to supply their own dial functions (return a user-controlled file descriptor) in the future.

Related to #150.