Add title_case_headers option (#218)
This required a custom HTTP mocker because no existing HTTP mocking library, including Mockito, don't allow you to access the literal header characters received since the HTTP spec says header keys are case-insensitive. This mocker will be tweaked and improved in a follow-up PR as there is also another set of tests that also need it.

Related to #205.