Replace middleware API with interceptors API (#206)
Implement an interceptors API and remove the old middleware API.

The new interceptors API is gated behind the `unstable-interceptors` feature flag. We sort of need to rush this out the door in order to implement #215 properly, but we're not ready to finalize the design yet. There are still some open problems to fix with this:

- `from_fn` doesn't work as desired. The trait bounds are too ambiguous for the compiler to infer for closures, and since the return type is generic over a lifetime, there's no way to give the return type the correct name using current Rust syntax.
- `InterceptorObj` wraps the returned future in an extra box.
- If an interceptor returns a custom error, it is stringified and wrapped in `Error::Curl`. We should introduce a new error variant that boxes the error and also records the type of the interceptor that created the error for visibility.
- Automatic redirect following currently bypasses interceptors for subsequent requests. This will be fixed when redirect handling is rewritten as an interceptor itself. See #232.

Users who are eager for interceptors can also opt-in to this feature and start experimenting if they can't wait for stabilization.

Related: #42