Fix timeouts causing unexplained EOFs
If a request errors after the response headers have already been received, but the full response body has not been read yet, then the response body stream will begin returning EOF without any other explanation other than what is written to the logger. This gives the appearance that everything is fine but half of the response body is missing.

To fix this, be sure to communicate to the response stream when a transfer error occurs while streaming the response body, so that the next call to read from the response will return an appropriate error instead of an unexplained EOF.

This is especially an obvious problem when combined with the timeout option, which can very likely occur during reading of the response body, whereas most other errors are likely to occur before that time.

See #154.