Improve documentation on timeouts (#249)
How timeouts are applied was a bit vague in our documentation and was causing some confusion. Add more detail to make it hopefully more clear as to what the timeout does, in particular that the timeout is still active even once the response headers are received and while the body has not been yet read.