Trim some heavy dependencies
Tis the season to be trimming the tree! Reducing dependencies where it makes sense is a regular task for Isahc to ensure shorter compile times, binary size, or package size.

This time around the goal was to eliminate `futures-util` from the essential dependencies, since it has long compile times. This included swapping out for `futures-lite` where needed, or just using our own code, as well as updating Sluice to remove `futures-util` from there as well.

Also a goal was to unify our preferred channel implementation, since we currently use _both_ `futures-channel` and `crossbeam-channel`, depending on the scenario. Here we can replace both with `flume` which has the nice property of supporting both sync and async operations on one channel.