Fix body reporting compressed length with auto decompression
When automatic decompression is enabled and a server returns a `Content-Length` response header for compressed content, `Body::len` was naively returning the unmodified value of `Content-Length` as the body length. This is incorrect because the `Content-Length` in this situation indicates the _compressed_ size of the body, not the uncompressed size.

Since `Body::len` is supposed to be reporting the size of the body that it is returning if known, returning the uncompressed size is confusing and buggy, and can (and does) cause issues for downstream users.

This fixes the issue by simply returning `None` as the body length whenever we are decompressing the response body on the user's behalf, since it is impossible for us to know the uncompressed size ahead of time.

Fixes #265.