Apply rustfmt to entire repo (#277)
Apply newly added rustfmt configuration from #276 to the entire repo. This will be the start of a clean slate going forward for consistent formatting.

In addition, this PR is to test semi-automated formatting with the `@teto-bot rustfmt` command, which can be invoked from any PR by the PR author or by an Isahc maintainer. Indeed, the formatting proposed by this PR was all automated.