Fix response body abort warning
Previously we were using two different atomic sources to discover when a response has been dropped, but only emitting a warning for an unconsumed response body if one of the sources indicated a drop. This was causing inconsistent results of when the warning would be emitted. Fix this by having just one source of truth for when the response is dropped.

In addition, change the warning log into an info log instead with a better message. Aborting a response stream is not incorrect and may be intentionally desired by the user. After #257 we could directly suggest using `consume()` in the log message.

See #270.