Always use a single space after colon for HTTP 1 headers
Previously headers were sent with no whitespace before or after the colon (`:`). This is legal according to RFC 7230, section 3.2, but is somewhat unusual and can cause poorly-written servers to choke on parsing the request. Update the encoder to always add a single space after the colon to be in line with what most parsers should expect.

This has no effect on HTTP/2 or newer, since they use a binary format for headers.

Fixes #286.