Add consume API (#284)
Add `ReadResponseExt::consume` and `AsyncReadResponseExt::consume` methods for reading and discarding a response body as described in #257.

Also update future types used by `AsyncReadResponseExt` to be `Send` whenever `T` is `Send` as appropriate. Consolidate the newtype futures with a macro that handles conditional `Send` properly to reduce boilerplate and chance of mistakes.

Fixes #257.