Refactor request configuration internal representation (#292)
Refactor how request configuration is represented internally from a purely extension-based model into a centralized struct. This reduces the generalized code somewhat, but makes it easier to understand for people less familiar with the codebase.

More importantly, this delivers a 5-10% performance improvement in synthetic benchmarks, since it greatly reduces the number of wasteful hashmap lookups we do when trying to apply request configuration to an easy handle.