Improve connection reuse log message (#335)
Improve the log message emitted when a connection is closed without being fully consumed. The message now includes a little bit more detail and links to [a new wiki page](https://github.com/sagebind/isahc/wiki/Connection-Reuse) that explains what is going on in more depth.

In addition, the message is no longer emitted for HTTP/2 or HTTP/3 connections (since it is not a problem there), nor if the connection cache is disabled (since _all_ connections are closed anyway).

See #334.